Fixed a bug where an error was logged that did not exist.
If the program was closed without being connected to the server, the program tried to disconnect from the server, which obviously failed.

+++ LiveFeedback.Desktop/Services/ServerService.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Threading.Tasks;
+using LiveFeedback.Models;
 using LiveFeedback.Shared;
 using LiveFeedback.Shared.Enums;
 using Microsoft.Extensions.Logging;
@@ -40,6 +41,9 @@ public class ServerService(
 
     public async Task StopServerAsync()
     {
+        if (_appState.ServerState != ServerState.Running)
+            return;
+
         try
         {
             await _signalRService.DeleteLecture(_appState.LectureId);
